+++ data-raw/reset_gauging_data.R
@@ -22,7 +22,7 @@ source("R/hyd1d-internal.R")
 
 ### open the connection using user, password, etc., as
 credentials <- credentials("/home/WeberA/hyd1d/DB_credentials_gauging_data")
-con <- dbConnect("PostgreSQL", 
+con <- dbConnect(drv = DBI::dbDriver("PostgreSQL"), 
                  host = credentials["host"], 
                  dbname = credentials["dbname"], 
                  user = credentials["user"],